+++ src/main/java/org/example/kunuz/controller/CategoryController.java
@@ -1,9 +1,13 @@
 package org.example.kunuz.controller;
 
 import org.example.kunuz.dto.CategoryDTO;
+import org.example.kunuz.dto.JwtDTO;
+import org.example.kunuz.enums.ProfileRole;
 import org.example.kunuz.service.CategorySevice;
+import org.example.kunuz.util.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageImpl;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,26 +20,46 @@ public class CategoryController {
     private CategorySevice categorySevice;
 
     @PostMapping("/create")// Category Yaratish
-    public ResponseEntity<CategoryDTO> create(@RequestBody CategoryDTO dto){
+    public ResponseEntity<CategoryDTO> create(@RequestBody CategoryDTO dto,
+                                              @RequestHeader(value = "Authorization") String jwt){
+        JwtDTO jwtDTO = JWTUtil.decode(jwt);
+        if (!jwtDTO.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         CategoryDTO result =  categorySevice.create(dto);
         return ResponseEntity.ok(result);
     }
 
     @PutMapping("/update/{id}")
     public ResponseEntity<CategoryDTO> updateById(@RequestBody CategoryDTO dto,
-                                                  @PathVariable("id") Integer id){
+                                                  @PathVariable("id") Integer id,
+                                                  @RequestHeader(value = "Authorization") String jwt){
+        JwtDTO jwtDTO = JWTUtil.decode(jwt);
+        if (!jwtDTO.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(categorySevice.updateById(id,dto));
     }
 
     @DeleteMapping("/{id}")
-    public Boolean deleteById(@PathVariable("id") Integer id){
+    public Boolean deleteById(@PathVariable("id") Integer id,
+                              @RequestHeader(value = "Authorization") String jwt){
+        JwtDTO jwtDTO = JWTUtil.decode(jwt);
+        if (!jwtDTO.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build().hasBody();
+        }
         categorySevice.deleteById(id);
         return true;
     }
 
     @GetMapping("/all")
     public ResponseEntity<PageImpl<CategoryDTO>> getAll(@RequestParam(value = "page", defaultValue = "1") Integer page,
-                                                        @RequestParam(value = "size", defaultValue = "10") Integer size){
+                                                        @RequestParam(value = "size", defaultValue = "10") Integer size,
+                                                        @RequestHeader(value = "Authorization") String jwt){
+        JwtDTO jwtDTO = JWTUtil.decode(jwt);
+        if (!jwtDTO.getRole().equals(ProfileRole.ADMIN)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok(categorySevice.getAll(page,size));
 
     }
